refactor(PostThread): move salon/thread_sidebar into ThreadSidebar/salon
