+++ frontend/core/unit/PostThread/ThreadSidebar/CommunityBrief.tsx
@@ -11,7 +11,7 @@ import useCommunity from '~/stores/community/hooks'
 import { assetSrc } from '~/utils/helper'
 import ImgFallback from '~/widgets/ImgFallback'
 
-import useSalon from '../salon/thread_sidebar/community_brief'
+import useSalon from './salon/community_brief'
 
 export default function CommunityBrief() {
   const { logo, title, meta, subscribersCount } = useCommunity()
